refactor: cleanup root namespace
