+++ src/WeihanLi.Common/Helpers/ProfilerHelper.cs
@@ -1,4 +1,5 @@
 ﻿using System.Diagnostics;
+using WeihanLi.Common.Services;
 
 namespace WeihanLi.Common.Helpers;
